Run Model Jags

Run.mod() takes a data frame and adjustable mcmc inputs to runjags::run.jags() as an mcmc bayesian model to estimate antibody dynamic curve parameters. The rjags::jags.model() was adapted from Teunis et al. (2016) modeling seroresponse dynamics to an infection. The antibody dynamic curve includes the following parameters:

y0 = baseline antibody concentration

y1 = peak antibody concentration

t1 = time to peak

r = shape parameter

alpha = decay rate

run_mod(
  data,
  nchain = 4,
  nadapt = 0,
  nburn = 0,
  nmc = 100,
  niter = 100,
  strat = NA
)

Arguments

data

A base::data.frame() with the following columns

nchain

An integer between 1 and 4 that specifies the number of mcmc chains to be run per jags model.

nadapt

An integer specifying the number of adaptations per chain.

nburn

An integer specifying the number of burn ins before sampling.

nmc

An integer specifying

niter

An integer specifying number of iterations.

strat

A string specifying the stratification variable.

name

description

Value

An jags.post list() object or multiple jags.post list() if jags is stratified, as well as a base::data.frame() titled curve_params that contains the posterior distribution. The jags.post is returned as a list() of class runjags::runjags. A curve_params base::data.frame() will also be exported with the following attributes

iteration = number of sampling iterations.

chain = number of mcmc chains run; between 1 and 4.

indexid = "newperson", indicating posterior distribution.

antigen_iso = antibody/antigen type combination being evaluated

alpha = posterior estimate of decay rate

r = posterior estimate of shape parameter

t1 = posterior estimate of time to peak

y0 = posterior estimate of baseline antibody concentration

y1 = posterior estimate of peak antibody concentration

stratified variable = the variable that jags was stratified by @export

References

Teunis PF, van Eijkeren JC, de Graaf WF, Marinović AB, Kretzschmar ME. Linking the seroresponse to infection to within-host heterogeneity in antibody production. Epidemics. 2016 Sep;16:33-9. doi: 10.1016/j.epidem.2016.04.001. Epub 2016 Apr 28. PMID: 27663789.

Author

Sam Schildhauer

Examples